requirements.txt
absl-py==1.2.0
alembic==1.6.5
anyio==3.6.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
autopep8==1.7.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
beniget==0.4.1
bleach==2.0.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
click==8.0.1
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
dominate==2.6.0
duplicity==0.8.21
email-validator==1.1.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
fasteners==0.14.1
fastjsonschema==2.16.1
filelock==3.8.0
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-Moment==1.0.1
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flatbuffers==1.12
fonttools==4.37.1
future==0.18.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg===1.16.0-unknown
greenlet==1.1.0
grpcio==1.47.0
h5py==3.7.0
html5lib==0.999999999
httplib2==0.20.2
idna==2.10
imageio==2.21.2
importlib-metadata==4.6.4
imutils==0.5.4
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.10
jsonschema==4.14.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
jupyterlab_server==2.15.1
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.4
langdetect==1.0.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
llvmlite==0.38.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.4
Markdown==2.6.8
MarkupSafe==2.0.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
natsort==8.0.2
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.11.0
networkx==2.8.6
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.2
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pbr==5.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
ply==3.11
progressbar2==4.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.12.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycups==2.0.1
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.18.0
python-editor==1.0.4
python-utils==3.3.3
pythran==0.10.0
pytz==2021.1
PyWavelets==1.3.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
screen-resolution-extra==0.0.0
seaborn==0.12.0
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.20
stack-data==0.5.0
stevedore==4.0.0
systemd-python==234
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.1.1
toml==0.10.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.6
usb-creator==0.3.7
virtualenv==20.16.3
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
visitor==0.1.3
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
Werkzeug==2.0.1
widgetsnbextension==4.0.2
wrapt==1.14.1
WTForms==2.3.3
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0alembic==1.6.5
Babel==2.9.1
blinker==1.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-Moment==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
langdetect==1.0.9
Mako==1.1.4
MarkupSafe==2.0.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.20
urllib3==1.26.6
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
